#104 [가이드화면] - 에러처리

+++ app/src/main/java/com/thequietz/travelog/record/view/RecordViewOneFragment.kt
@@ -98,9 +98,6 @@ class RecordViewOneFragment : Fragment() {
                 if (recordViewOneViewModel.isCommentChanged(currentText) && currentText != "") {
                     showChangeCommentDialog(currentText)
                 }
-                binding.etRecordViewOne.isEnabled = !binding.etRecordViewOne.isEnabled
-                binding.tvRecordViewOneSave.visibility = View.GONE
-                binding.ibRecordViewOneEditComment.visibility = View.VISIBLE
             }
             tvRecordViewOneReduce.setOnClickListener {
                 val popup = PopupMenu(requireContext(), it)
@@ -131,6 +128,11 @@ class RecordViewOneFragment : Fragment() {
             .setNegativeButton("예") { dialog, which ->
                 CoroutineScope(Dispatchers.IO).launch {
                     recordViewOneViewModel.updateComment(currentText)
+                    CoroutineScope(Dispatchers.Main).launch {
+                        binding.etRecordViewOne.isEnabled = !binding.etRecordViewOne.isEnabled
+                        binding.tvRecordViewOneSave.visibility = View.GONE
+                        binding.ibRecordViewOneEditComment.visibility = View.VISIBLE
+                    }
                 }
                 Toast.makeText(requireContext(), "현재 내용이 저장되었습니다", Toast.LENGTH_SHORT)
                     .show()
@@ -144,11 +146,24 @@ class RecordViewOneFragment : Fragment() {
             .setTitle("삭제 확인")
             .setMessage("현재 이미지를 삭제하시겠습니까?")
             .setNegativeButton("예") { dialog, which ->
-                CoroutineScope(Dispatchers.IO).launch {
-                    recordViewOneViewModel.delete()
+                if (recordViewOneViewModel.currentImage.value?.id != null) {
+                    CoroutineScope(Dispatchers.IO).launch {
+                        recordViewOneViewModel.delete()
+                    }
+                    Snackbar.make(
+                        binding.layoutRecordViewOne,
+                        "이미지가 삭제되었습니다",
+                        Snackbar.LENGTH_SHORT
+                    )
+                        .show()
+                } else {
+                    Snackbar.make(
+                        binding.layoutRecordViewOne,
+                        "삭제할 이미지가 없습니다",
+                        Snackbar.LENGTH_SHORT
+                    )
+                        .show()
                 }
-                Snackbar.make(binding.layoutRecordViewOne, "이미지가 삭제되었습니다", Snackbar.LENGTH_SHORT)
-                    .show()
             }
             .setPositiveButton("아니오") { dialog, which ->
             }.show()
